requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
anyio==3.7.1
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.5.14
decorator==5.1.1
deprecation==2.1.0
dnspython==2.4.2
docx2txt==0.8
duckdb==0.8.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
extra-streamlit-components==0.1.60
fastapi==0.103.2
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
googletrans==4.0.0rc1
greenlet==2.0.2
grpcio==1.59.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httptools==0.6.0
httpx==0.13.3
huggingface-hub==0.16.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.334

langsmith==0.0.63

loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numexpr==2.8.7
numpy==1.24.4
onnxruntime==1.16.0
openai==0.27.10
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pandas==1.5.3
pandasai==1.4.8
parso==0.8.3
Pillow==9.5.0
pinecone-client==2.2.2
polib==1.2.0
posthog==3.0.2
prompt-toolkit==3.0.40
protobuf==4.24.3
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==1.10.13
pydeck==0.8.0
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.1.1
pypdf==3.16.2
PyPika==0.48.9
PyQt6==6.6.0
PyQt6-Qt6==6.6.0
PyQt6-sip==13.6.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
rich==13.5.3
rpds-py==0.10.3
scipy==1.11.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
stack-data==0.6.3
starlette==0.27.0
streamlit==1.26.0
streamlit-authenticator==0.2.3
streamlit-chat==0.1.1
streamlit-lottie==0.0.5
streamlit-modal==0.1.0
streamlit-option-menu==0.3.6
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
tk==0.1.0
tokenizers==0.14.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.9
websockets==11.0.3
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.17.0
